Refactor SCTP and SCTPTransmitter for improved state management and timer handling
